Made particles work!
And delta is now a float, so much larger velocities is needed now.
The current entities is adapted to the new delta.

Multiply old vels with 1000

+++ core/src/se/doverfelt/effects/Effect.java
@@ -6,7 +6,7 @@ import se.doverfelt.PongzStart;
  * Created by rickard on 2016-02-03.
  */
 public interface Effect {
-    void update(PongzStart world, int delta);
+    void update(PongzStart world, float delta);
     void create(PongzStart world, String name);
     String getName();
     String getEffectType();
